pala

+++ 0x03-python-data_structures/linked_lists.c
@@ -7,7 +7,7 @@
  * @h: pointer to head of list
  * Return: number of nodes
  */
-int print_listint(const listint_t *h)
+size_t print_listint(const listint_t *h)
 {
     const listint_t *current;
     unsigned int n; /* number of nodes */
